requirements.txt
pandas
matplotlib
shodan
beautifulsoup4
python-dotenv
transformers
torch
scikit-learn
argostranslate
hf_xet